refactor: change return type to size_t for cross section vertex/contour count (#1106)

+++ bindings/c/cross.cpp
@@ -213,11 +213,11 @@ double manifold_cross_section_area(ManifoldCrossSection *cs) {
   return from_c(cs)->Area();
 }
 
-int manifold_cross_section_num_vert(ManifoldCrossSection *cs) {
+size_t manifold_cross_section_num_vert(ManifoldCrossSection *cs) {
   return from_c(cs)->NumVert();
 }
 
-int manifold_cross_section_num_contour(ManifoldCrossSection *cs) {
+size_t manifold_cross_section_num_contour(ManifoldCrossSection *cs) {
   return from_c(cs)->NumContour();
 }
